+++ src/lexer.cpp
@@ -75,7 +75,14 @@ auto Lexer::generate_token() -> std::optional<Token> {
     case '}': consume(); return Token{"}", line_, col_ - 1, col_ - 1, TokenType::CLOSE_CURLY};
     case '[': consume(); return Token{"[", line_, col_ - 1, col_ - 1, TokenType::OPEN_SQUARE};
     case ']': consume(); return Token{"]", line_, col_ - 1, col_ - 1, TokenType::CLOSE_SQUARE};
-    case ':': consume(); return Token{":", line_, col_ - 1, col_ - 1, TokenType::COLON};
+    case ':': {
+        consume();
+        if (peek(':')) {
+            consume();
+            return Token{"::", line_, col_ - 2, col_ - 1, TokenType::DOUBLE_COLON};
+        }
+        return Token{":", line_, col_ - 1, col_ - 1, TokenType::COLON};
+    }
     case ';': consume(); return Token{";", line_, col_ - 1, col_ - 1, TokenType::SEMICOLON};
     case '(': consume(); return Token{"(", line_, col_ - 1, col_ - 1, TokenType::OPEN_BRACKET};
     case ')': consume(); return Token{")", line_, col_ - 1, col_ - 1, TokenType::CLOSE_BRACKET};