+++ src/visitor.hpp
@@ -6,6 +6,7 @@
 class ParaDecl;
 class LocalVarDecl;
 class GlobalVarDecl;
+class EnumDecl;
 class Function;
 class EmptyExpr;
 class AssignmentExpr;
@@ -22,6 +23,7 @@ class CallExpr;
 class CastExpr;
 class ArrayInitExpr;
 class ArrayIndexExpr;
+class EnumAccessExpr;
 class EmptyStmt;
 class CompoundStmt;
 class LocalVarStmt;
@@ -37,6 +39,7 @@ class Visitor {
     virtual auto visit_para_decl(std::shared_ptr<ParaDecl> para_decl) -> void = 0;
     virtual auto visit_local_var_decl(std::shared_ptr<LocalVarDecl> local_var_decl) -> void = 0;
     virtual auto visit_global_var_decl(std::shared_ptr<GlobalVarDecl> global_var_decl) -> void = 0;
+    virtual auto visit_enum_decl(std::shared_ptr<EnumDecl> enum_decl) -> void = 0;
     virtual auto visit_function(std::shared_ptr<Function> function) -> void = 0;
     virtual auto visit_extern(std::shared_ptr<Extern> extern_) -> void = 0;
     virtual auto visit_empty_expr(std::shared_ptr<EmptyExpr> empty_expr) -> void = 0;
@@ -54,6 +57,7 @@ class Visitor {
     virtual auto visit_cast_expr(std::shared_ptr<CastExpr> cast_expr) -> void = 0;
     virtual auto visit_array_init_expr(std::shared_ptr<ArrayInitExpr> array_init_expr) -> void = 0;
     virtual auto visit_array_index_expr(std::shared_ptr<ArrayIndexExpr> array_index_expr) -> void = 0;
+    virtual auto visit_enum_access_expr(std::shared_ptr<EnumAccessExpr> enum_access_expr) -> void = 0;
 
     virtual auto visit_empty_stmt(std::shared_ptr<EmptyStmt> empty_stmt) -> void = 0;
     virtual auto visit_compound_stmt(std::shared_ptr<CompoundStmt> compound_stmt) -> void = 0;